update  orders,and delete Order

+++ database/src/main/java/by/itacademy/boldysh/database/dto/AdditionalServiceDto.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class AdditionalServiceDto {
 
     private String services;
-    private Integer cost;
+    private BigDecimal cost;
 
 }
