+++ database/src/main/java/by/itacademy/boldysh/database/entity/OrderRentalCar.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -30,7 +31,7 @@ public class OrderRentalCar extends BaseEntity<Long> {
     private LocalDate dateFinishRental;
 
     @Column(name = "cost")
-    private Integer cost;
+    private BigDecimal cost;
 
     @Column(name = "status_order")
     @Enumerated(EnumType.STRING)
@@ -40,7 +41,7 @@ public class OrderRentalCar extends BaseEntity<Long> {
     @JoinColumn(name = "additional_services_id")
     private AdditionalService additionalService;
 
-    public OrderRentalCar(Long userServiceId, Long carId, LocalDate dateStartRental, LocalDate dateFinishRental, Integer cost, StatusOrder statusOrder, AdditionalService additionalService) {
+    public OrderRentalCar(Long userServiceId, Long carId, LocalDate dateStartRental, LocalDate dateFinishRental, BigDecimal cost, StatusOrder statusOrder, AdditionalService additionalService) {
         this.userServiceId = userServiceId;
         this.carId = carId;
         this.dateStartRental = dateStartRental;
