+++ service/src/main/java/by/itacademy/boldysh/service/interfaces/CarService.java
@@ -6,10 +6,12 @@ import by.itacademy.boldysh.service.util.BaseMethod;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
+
 
 public interface CarService extends BaseMethod<Long, Car> {
 
-    void updateCar(Car car, Integer costRentalOfDay);
+    void updateCar(Car car, BigDecimal costRentalOfDay);
 
     Page<Car> findByFilterAndPaginationCars(FilterDto filterDto, Pageable page);
 }
